Fix Mobility QtDeclarative plugins builds
"target" is already in INSTALLS, so don't add it again. Doing so
causes tons of make warnings about overriding install and uninstall
targets.

Don't set DESTDIR to Qt's dir when compiling. Let the libraries stay
where they are here, but install to Qt only.

Note: installation should go to $PREFIX/imports, not to
$$[QT_INSTALL_IMPORTS]. Installing outside $PREFIX is a bad idea.

Reviewed-By: Ossi