Add IPC failure error reporting
Errors can occur in IPC and cannot be silently ignored.
These are now reported through a signal from the client
service object:

errorUnrecoverableIPCFault(QService::UnrecoverableIPCError)

and should be connected to by the client to receive error
reports.  Services must not use this name, and argument
functions as to avoid having their function masked by
this signal.