Cleaned up some structural cruft with Maps
QGeoTiledMappingManagerEngine used to contain a map of
QGeoTiledMapData objects to "handler" objects.  This was done when
QGeoTiledMapData was more lightweight and did not inherit from QObject.

Since that's not the case at the moment I folded the "handler"
functionality back into QGeoTiledMapData, which avoids needing to
pass a lot of data back and forth between a data object and an object
in a one-to-one mapping with that data object.