Improved the fix for QTMOBILITY-311
The previous fix works and would report positions to the user, but
since the timestamps were absent from positions from non-satellite
methods QGeoPositionInfo::isValid() would have returned false.

This is a fix for that, and adds some comments describing the
outcome of some logging which will be helpful when doing further work
on the code.