Include QtMultimediaKit moc files in .cpp files.
Because the QT namespace is used instead of the QTMOBILITY namespace
this library is the only one in QtMobility not to do this.

This is also the only library for which Symbian builds are failing with
moc related errors.  Hopefully removing one point of differentition will
also remove the other.