Parse ADR details differently in maemo5 backend
We now only parse the first 7 fields of the ADR attribute type, which
are supported by the maemo5 backend.  We do not qFatal() or qCritical()
if the number of fields is more or less than we expected, but instead
attempt to deal with the difference gracefully.
This relates to task QTMOBILITY-380.

Also, fix a crash introduced by ee0e4456d56165d when the user attempts
to retrieve a nonexistent contact.