Fix: adding a service sometimes fails with UnknownError code
We need to make sure that when service ids are generated
that we reseed the random number generator.  Sometimes
a Qt application on *nix platforms will be automatically
seeded to 1 if qrand() is called before srand().
Because Uuid on *nix platforms depends on the rand()
we need to make sure reseeding happens.

Task-number: QTMOBILITY-388
Reviewed-by: lorn potter