Patch from simulator team:
The Simulator wants Mobility Messaging to always use the
QMF backend as it allows easy inter-process synchronization
between Simulator and Messaging-using clients.

As such, the main change is to the build system: Win32 and Mac
also want to use the QMF plugin.

The second change is to the QMF backend: We've wrapped all calls
to QMailStore::instance() to be able to adjust the QMF_DATA
path before the backend is initialized.