Modified the map objects to use the Graphics View system internally.
This avoids us having to implement, test and maintain code with very
similar functionality to the code that has already been implemented
and is being tested and maintained as part of the Graphics View system.

There have been some concerns raised about the efficiency of Route
objects while using the Graphics View system.

These were addressed by filtering the lines that make up the route
based on a minimum manhattan distance for disaply (which was already
done in the old system) and by creating a QGraphicsPathItem only for
the section of the route that is visible.