Fix flicker in video widget on windows.
We need to disable updates when using the window control because the
video producer writes directly to the window, and Qt painting the
(black) back buffer to the window will wipe that out.

The also prevents us from receiving paint events so repaint on WM_PAINT
window message instead.

Finally we also need to fill the background if the video is not
currently being rendered.