Start implementing notifications beginning with category notifications
The implementation is based on using a QFileWatcher on the database
file.  Inside the database we also have two new tables
landmark_notifications and category_notifications.  These
tables are essentially logs for added/modified/removed landmarks
and categories.  Whenever a database change is detected, we check the
tables to see what has changed.

Whenever a new entry is made to one of these notification tables,
all entries that are more than two minutes old are removed.  This
should be sufficient time for any notifications to be made whilst
not letting the notification tables become too large.