Test category notifications + refactor sqlite tests
The sqlite tests have been refactored so that the slots
are no longer implemented as inline functions.  This allows for
easier navigation since we can use the class definition as an index.

The importLandmarkGpx function has been disable temporarily for now.
When we save landmarks, it is currently implemented as saving
them one at a time.  Each save is causing a database change, which
is triggering the filewatcher on the database and thus we are
being flooded with notification changes.