Change the result of a gallery query request to a result set interface.
The major difference between the two is in the list theoretically all
items are accessible simultaneously while in the result set only
one item is available at a time and it's necessary to seek to access
another.

A number of suspect features are removed with this change as well, they
include thumbnail loading, and the Item, Url, and Count requests.
Thumbnail loading and the Count request will likely return in another
form, while any use case addressed by the Item, or Url request should
be better addressed by the Query request already.