Improved behaviour of tiled maps when viewing the date line.
The code related to this was failing in several different ways and
most of these failures have now been addressed.

Some of the problems came from QGeoBoundingBox, which requires
further fixes.

There are still some glitches with the clearing of tiles from the
zoom cache and the handling of the zoom preview.

The drawing of the rectangle map objects was modified so that they
can be drawn across the date line.  This will be more difficult for the
other objects.