Expand the documentation of QContactThumbnail and QContactAvatar
The Avatar detail contains URLs of avatars; the resource which a URL
points to might be owned by someone else, and so that person can
modify the content of that resource at will.  A thumbnail is static
data which is set by the owner of the contact, rather than the person,
group or organization which is represented by the contact.