Remove platform attributes.
We need to think more deeply about what benefits the platform
attributes give.  So far it would enable searching addresses on s60
but instead we could make it easier by providing QLandmarkAddressFilter

More time is needed to consider the use cases.  In future we may
have a QLandmarkFreeText filter which would search names, address,
category etc.  So for now we are not including platform attributes.