Fix 3.1 sensor backend packaging.
The 3.1 sensor backend now has the same name as the 3.2+ sensor backend.
This is fine because only 1 backend can run on a given device.

When creating a package that should run on 3.1 and 3.2+ devices,
EPOCROOT31 should point to a 3.1 SDK (that will contain the 3.1 backend)
while EPOCROOT should point to a 3.2+ SDK (that will contain the 3.2+ backend).

It is possible to build the 3.1 backend on a 3.2+ SDK with a bit of work.
1) Install the 3.1 sensor API plugin (RRSensorAPI) onto the 3.2+ SDK.
2) Build the 3.1 sensor backend.
3) Copy the 3.1 sensor backend and the generic sensor backend into another directory.
4) Build the 3.2+ sensor backend.
5) Set EPOCROOT31 so that the 3.1 sensor backend can be found.
6) Run qmake in src/s60installs and build the sis file.

Fixes: QTMOBILITY-423