Refactor database operations
Expose saveLandmarkHelper and saveCategoryHelper in the database
database operations namespace.  The helper function perform a save
operation without any transaction handling, which is left up to the
caller.  We do this specifically for the case of importing,
we don't want to save categories or landmarks unless all
can be saved and to do this we need to do it in one transaction.
Previously each save category was a single transaction.