Implement import of categories using lmx
Add unit tests.
Also ensure a category with an empty name cannot be added.
The QLandmarkFileHandler lmx is now more tightly coupled
with the database operations.  This is because we need to add
categories if they already do not exist.  At the same time we
want to do the import operation atomically and to do so we
need to control the transaction.