Some further improvements to the map objects code.
The QGeoTiledMapObject should be able to be created in a plugin based
on the public API.  The previous version of QGeoMapObjectInfo -
which stores the QGeoMapData subclass specific code for each of the
QGeoMapObject subclasses - were not able to be used via the public API.

This commit is primarily about moving towards an API in which the
QGeoTiledMapData class could have been developed in the plugin code,
which involved some changes to the map data and map object classes.

When I've slept and caffeinated more (in that order) I'll double check
that the API is still in a sane state.

Also:
- added preliminary support for map overlays
- removed the navigator class since it won't be in the first release
- various other little cleanups

 QGeoMapObjectInfo