Remove QLandmarkFetchHint, replace with limit and offset vars
The QLandmarkFetchHint wasn't really a good place to put the limit
and offset since it implies they can be ignored.  The original
intention of the fetch hint was to have a place to put hints
for partial landmark retrieval.  This wasn't and explicit
use case placed upon us so we technically don't need it.

If we need to have partial landmark retrieval we could always
add it to the QLandmarkFetchRequest later.