Remove QLandmarkDistanceSort
The distance sort was deemed to have too much of a performance impact.
While a convenient and useful class, its utility is limited
if it will not be performant.  Even if we do says such sort order
has emulated functionality, people may try to use it and be severely
disappointed, which is what we don't want.