Add a light sensor.
As opposed to the ambient light sensor, the light sensor
reports in lux. It also indicates the field of view of the
sensor.

As the N900 can report lux values it reports via this API.
Legacy compatibility is provided by the generic ambient light sensor
which uses the logic previously used by the N900 ambient light sensor
to get ambient light enum values from lux values.

Fixes: MOBILITY-1268