Symbian: Refactored recurrence rule conversions.
Now most of the conversions are not dependent on the recurrence rule
frequency. This should give better flexibility to the backend if
symbian calendar server will support some of the currently
unsupported recurrence parameter combinations. Also the test
cases were changed to give a warning for test cases that are not
currently supported.