Document Sensors.conf
As part of this change, the system will no longer consider
~/.config/Nokia/Sensors.conf. Only /etc/xdg/Nokia/Sensors.conf
will be considered. Since Sensors.conf was not previously
documented and since it is designed for the system to specify
defaults (rather than for users to override defaults) this
is considered a safe change.

Note that a mobility build with tests enabled will still look
at ~/.config/Nokia/Sensors.conf because the unit test needs
to modify it but may not be able to access
/etc/xdg/Nokia/Sensors.conf.

Fixes: QTMOBILITY-441