Fix behaviour of fetching, saving and removing requests
When fetching by ids, the resultant number of
landmarks should be the same as the number of ids.
In cases where a landmark could not be retrieved
for a given id, a default constructed landmark
should be returned.

This is the same case for saving and removing landmarks.
When saving landmarks.  For new landmarks, any which
could not be saved are untouched while those successfully
saved will be assigned a new id.  For pre-existing landmarks
the landmarks in given to the save operation should
effectively remain untouched.  Likewise for landmark removal
the list of ids given to the operation should remain untouched.
We need to do this because the error map will be used
to index into the list of landmarks/landmark ids and they
need to match up.