Fixed bug in QGeoCoordinate comparison.
Coordinates at the poles were considered to be unequal when they
had different latitudes.

Also cleaned up some documentation and made a small change to the
distanceTo function to improve its numerical stability for large
distances.  The maths is equivalent - the change just uses asin
instead of atan2 to avoid some edge effects - and the tests still seem
to pass.