Fix deletion of requests
When the sqlite backend updates a request it does so via the queued
connection so that if the operation was finished, an update  gets
queued.  But during this time the request could have been deleted
a new request made with the same address.  By the time the updates gets
dispatched, it tries to update the new request since it has the same
address as the old.

To get around this we associate a runId with each request.  This runId
refers to the query run.  i.e. if a single request was started 3 times,
then it goes through 3 query runs, hence we have 3 run Ids.
When a query run is finished it or deleted it's query run id gets
removed.  The run ids are stored in m_activeRequestsRunIdHash.