Fix and refactor notifications
Ensure that we have millisecond resolution for the latest timestamps
when the  engine is created.  Otherwise there may be issues in
receiving extra notifications, eg engine creation is at 5.4 seconds;
because we had a resolution of seconds, the latest timestamp
was at 5 seconds.  This would pick up notifications at say,
5.3 seconds.