Give a default document type parameter to exporter functions
The QVersitContactExporter::exportContacts() and
QVersitOrganizerExporter::exportItems() functions previously had a
mandatory second parameter to specify the version of the output document.
A default value has now been given so the user doesn't need to specify
it.  This default value once existed in a pre-1.0 version of the API
but had to be removed because it conflicted with an older, identically-
named function (which no longer exists).