Expand documentation for various fields in QContactAddress
The documentation of QContactAddress detail was lacking in clarity.
This commit adds documentation which makes it clear what each field
is used for, based on the description given in RFC2426:
"The structured type value corresponds, in sequence, to the post
office box; the extended address; the street address; the locality
(e.g., city); the region (e.g., state or province); the postal code;
the country name."

Relates to task QTMOBILITY-426.