Removed start*Request functions from engineV2 interface
These sounded good at the time as a way to eliminate switch blocks from engines,
but engines generally just put all requests into a queue, regardless of type,
and work through the queue in a separate thread (at which point the switch
block is necessary).  These functions tend to just get in the way of engine
implementations.