Symbian: Implemented collection signalling
Emitting right signals proved difficult because of the fact that the calendar
file might not be removable. EMarkAsDelete flag is set instead. This causes a
heap of ugly checking.

Currently there is at least one bug where an extra changed signal is emitted in
a situation where the client is enabling a calendar which is marked for deletion.