Change the sensors unit test.
We can't have the sensors library using UserScope by default.
Instead there is a unit test shim function (which is exported but
not documented and not in any headers). The unit test calls this
shim function to switch various state bits in the library so that
the unit test can do what it needs to do without affecting the
default state for apps using the library.