Symbian: Workaround for calendar server UTC problem
The calendar/agenda server has a problem with UTC times. It will move
the UTC time with daylight saving time. UTC time should be fixed
always and the actual local time for a calendar entry should be
calculated from the UTC time. Fortunately local time's reported
from the server are correct and using it with TCalTime/QDateTime
conversions seems work as as workaround for this issue.