Fix crash in QGalleryTrackerResultSet.
Overlap the scan areas by doubling the size of the inner loop rather
than halving the steps of the outer loop.

This is because the end value of the outer loop is rounded up to the
nearest multiple of the step size, so incrementing in anything other
than the step size will potentially result in addressing items past
the end of the vector.