Implement QAudioInput::suspend() using CMMFDevSound::Stop()
As with playback mode, lack of support for CMMFDevSound::Pause() in
DevSound's recording mode causes problems on some devices.
Specifically, while QAudioInput works fine on the Nokia 5800, this
bug was observed on the Nokia N8.  This fix means that suspending
and resuming audio input will work correctly on all devices.

Task-number: QTBUG-13506
Reviewed-by: Derick Hawcroft