Remove QGalleryRemoveRequest from the QtGallery API.
This request is of little use in its current form as an document removed
from the database but not deleted will just be restored on a subsequent
scan.

Actually deleting the file would have been another option, but modifying
the file-system doesn't otherwise fall within the scope of the document
gallery so it's best not to make an exception here.