Remove examples/s60installs/s60installs.pro
No need for the s60installs.pro for examples, because
examples are targeted to be installed individually.

Examples use development UID's and they rely on data in
application private folder (named according to uid).
Therefore, if such combined package is needed in future, this
common package file needs to be created from scratch and
needed changes applied to example applications.

Reviewed-by: Gareth Pethig