Removed the last bit of code that is being put aside for later.
The code related to marker objects and their ordering has been
put aside into a separate branch pending a proper consideration
of what is required and how to achieve it with suitable API quality.

Several reverts prior to this were necessary to remove this
functionality from the beta - this change alters a small part of a
needed change that uses the removed functionality.