Allow building contacts by itself.
Logic in s60installs.pro did not match the logic for the directories
that actually build the stuff s60installs.pro packages. This meant
that contacts could not be built by itself.

As a fix, update the s60installs.pro logic to match the logic for
the things it packages.

Fixes: MOBILITY-1484