Fixes: MOBILITY-1428 landmarkbrowser missing functionality
Modify layout to be usable on N900.

Also fix race condition where progress dialog is shown after
a request has finished, meaning the dialog won't receive
a notification to be hidden, resulting in the progress dialog staying
shown.  We've worked around this by deleting the dialog
but there probably is a better way to handle it.