Updated QGeoRouteSegment to follow LBSP semantics.
The route segment know appears like a linked list.

We probably don't want to add the same semantics to the maneuver
class until there's a solid need for it, which would occur when
we add alternativeRouteSegments() alongside nextRouteSegment(),
ie when the segments become a graph with nextRouteSegment()
describing a canonical preferred path through the graph.