Fix bug in memory backend related to request deletion
If the client deletes the request in a slot connected to stateChanged()
the memory backend in startRequest() updates the request state to
active and then attempts to perform it.  This would result in problems
because the request had been deleted.  This patch fixes this problem
and modifies the unit test to prevent regression.