Proposal to modify the behaviour of items* methods to include start and end dates
QOrganizerItemManager::items now contain a start and end period, and will return the results between these dates.
Beside of that it will expand the occurances of the items if FindAllOccurances is specified as a parameter (it is the default behaviour),
otherwise it will include the originating parent node in the results if his occurances are between the given interval.