Fixed badly named autotests.
It's very important that TARGET and the test class name are matching,
which was not true for these tests.

If a test fails to compile, or crashes, or hangs, then we assume that
TARGET is the name of the testcase which would have been written to the
test log.  To make this assumption correct, we need to enforce this
naming convention.