Fix QAbstractVideoSurface assert when Qt version changes.
To allow QtMultimediaKit to build without access to Qt's private headers
and keep binary compatible with the QAbstractVideoSurface in Qt, we
included a copy of QObjectPrivate in qabstractvideosurface_p.h.  This
created a situation where QAbstractVideoSurface could not be run using
a different version of Qt to the one it was built against.

This removes the copy of QObjectPrivate, and instead uses QObject
dynamic properties to store member variables.

Reviewed-by: Justin McPherson