Symbian: Saving collection with only a name key in metadata
Previously to save a collection a filename was required. Now if a filename is not provided
it will use the name as a filename. This way the client needs only to provide a name for
the collection and not worry about the actual calendar filename.

Also added some more metadata verifying steps to tst_symbianomcollections::saveCollection test.
This uncovered some issues with timestamps. Fixed them as well.

Relates to MOBILITY-1606