Get maemo5 backend compiling with the API changes
Removal of reminder.dateTime() means we need to use the delta always,
similarly, removal of the locationAddress+Name replaced with just the
location label means we don't need to do concatentation to store it.
Also, the datatype of the geolocation parts has changed (from string
to double).  This commit fixes these issues for the maemo5 backend.