Fixes for various doc errors in Location.
I was unable to reproduce the error output on my system even after
updating my version of Qt, although both Peter and Alex were producing
similar lists of errors.  Because of this it's hard to know if this
commit fixes all of the doc errors or even if it has introduced new errors.

I've only focused on the errors in the positioning and maps and navigation
parts of the API since they're what I'm familiar with.