Potential fix for MOBILITY-1602
The problem was a segfault which was occuring because the map type
signals were connected before the map widget was created - if the creation
of the map widget failed then the app would crash.

This kind of failure had mostly been addressed previously by moving
most of the signal connection code to a function which was only called
once the widget had been setup successfully.

This patch moves the map type signal connection code into that
function.