Potential fix for MOBILITY-1669.
Changed the example to quit after reaching the unrecoverable
(and previously assumed to be rare) error message as mentioned in the \
bug report, since the example won't be able to function if it's
reaching that point in the code.

As an additional protection from unexpected behaviour, the "Download"
button is now connected after a network connection is established.