Refactor id comparison unit test, add empty item manipulation test
The id comparison unit test used some item and detail types which are
not supported on all engines.  This test has been refactored to use
only the common item types.

An empty item manipulation test has been added, to ensure that each
backend behaves sensibly when the inputs given to it are empty or
null.