Clean up the document properties example.
Add a simple binding class which maps a property in a result set to a
QObject property, instead of the horrid list of properties and label
names.

This also opens the door to editable properties and intelligent
selection of widgets for properties instead of trying to pick an
appropriate one based on type.