Enable image thumnbails in mediabrowser on symbian.
Use QThread instead of QtConcurrent to handle the loading of images
in mediabrowser.  QFuture isn't enabled on symbian so the old solution
had to be disabled.  Using QThread also means we can lower the priority
of the image loading thread which improves the responsiveness of the
UI on other platforms.

Task-number: MOBILITY-1816
Reviewed-by: Justin McPherson