Implement a vCard preserver plugin
This plugin will try to preserve a vCard on an import/export round-trip.

Also, moved the "restore" part of the backup plugin into the core importer
code.  This is because restore must happen before the preserver plugin runs.
It would've been possible to split the backup plugin into two plugins, but
that would've been unnecessary plugin-loading overhead.  This changes the
behaviour of the importer to always recognise backed-up data, which is
desirable anyway

Relates to MOBILITY-1410

Change-Id: I9c60c092e7fd85a1c2555f16f5c9c6be418e4ae9