Fixed some issues with QGeoMapGroupObject.
Removing a group object backed by QGeoTiledMapGroupObjectInfo was
causing a segfault.  The group info object was removing its graphic item
from the scene before the children attempted to remove themselves from the
scene, so the children were being removed twice.

This was fixed by having the QGeoTiledMapGroupObjecInfo destructor
removing the childrens graphic items, so that this occured first.

There were also issues with adding / removing the groups children to
the scene which were addressed by a few nearby changes.