Support for QGraphicsVideoItem in mediaplayer Symbian MMF backend
The 'overlay' implementation of QGraphicsVideoItem is used.  This
means that the QGraphicsView window handle is provided to the
backend, along with a QRect relative to the window.

The QGraphicsVideoItem::paint() function uses a color key set by
the backend to paint into the QGraphicsView; the MMF backend sets
this color to Qt::transparent.

The QGraphicsVideoItem implementation is modified to install an
event filter on the native widget into which video is rendered.
This is so that it can listen for the WinIdChange event which is
sent just before the native window handle is destroyed.  At this
point, the native window handle is removed from the MMF video
player object to prevent the MMF from later trying to access an
invalid window handle.  This depends on a patch in QtGui to
ensure that WinIdChange is emitted during native widget
destruction; the patch is tagged with QTMOBILITY-645.

A further change to the QGraphicsVideoItem implementation is
that, if the color key set by the backend is non-opaque (as in
this case), then QPainter::CompositionMode_Source is used when
painting the item.

Known limitations of this implementation:

* Because this implementation requires transparency to be
  painted into the backing store, Qt::WA_TranslucentBackground
  must be set on the application's TLW before it is shown.

* Because Symbian allows only one non-UI surface (e.g. a video
  surface) to be attached to a window, only one video clip can
  be rendered to any QGraphicsView.  If multiple
  QGraphicsVideoItems are added to the scene, only the last one
  will show video; the rest will just appear as black rectangles.

Task-number: QTMOBILITY-645
Reviewed-by: Antti
Reviewed-by: Dmytro Poplavskiy