Provide a blocking default implementation of waitForFinished().
The previous implementation of QGalleryAbstractResponse::waitForFinished
simply returned true, assuming that it would always be overridden when
a true implementation is required. Ideally it will still be overridden
but if it's not possible to better than blocking on the event loop then
we have a single fallback solution.

Task-number: MOBILITY-1861
Reviewed-by: Dmytro Poplavskiy