Versit: Fix a crash with older versions of gcc.
The code in question relies on the fact that binding a temporary to
this const reference should extend the temporary's lifetime to the
lifetime of the reference. However, this seems not to work correctly
with older versions of gcc: it fails with gcc 4.2.4 and works fine
with gcc 4.4.5.

Done-with: Niels Weber