Fix race condition in request-class destruction
Previously, the vtable of the request could be that of the base class
instead of the derived class during a small window of time during
which the engine could attempt to access some of the meta-object
properties.  This commit ensures that each request notifies the engine
of its destruction in its destructor rather than in the destructor
of the base class, which prevents this race condition.

Task-number: QTMOBILITY-711
Reviewed-by: Michael Goddard