Update contacts unit test given service framework integration
The sendemailaction is now only installed if mobility is configured
with the service framework module enabled.  This change moves the test
into the tst_qcontactaction test which is only enabled if the service
framework module is enabled, and updates the test so that it works
properly given the service framework integration changes.

Task-number: MOBILITY-1950
Reviewed-by: Michael Goddard