Remove some dead code from QVersitContactExporter.
At some point in the past, I removed the possibility of error from
QVersitContactExporterPrivate::exportContact.

However, I forgot to remove the error return parameter or the boolean
return value from the function.  This commit removes these because it
always returns true and it never uses the error parameter.  This also
squashes a compiler warning (unused parameter).

Change-Id: I15c4e7e1194f837c0d95409fb73bd6855ff9ae01