Ordered the location tests.
This is needed for "make check" to work reliably.

Previously we had tests with plugins and unit tests in the same folder,
but this caused trouble with make check.

The workaround was to create XYZplugins and XYZ as separate directories
in location.pro.  This was done, but I forget to enforce the
ordering of the entries in location.pro