Only allow adding alarms in calendardemo if supported by the backend
Previously, the UI elements to add an alarm were provided regardless
of whether or not the backend supported that detail.  This commit
checks to see whether the VisualReminder detail is supported for the
given item type, and if so, adds the UI elements, instead of adding
the UI elements regardless of the level of support.

Task-number: QTMOBILITY-746
Reviewed-by: Kevin Wu Won