Ensure that the display format of the date time edit widgets is set
Previously, no explicit display format for the date time edit widgets
was set, which could result in an invalid datetime being used as the
start or end date in a todo or event, causing the save operation on
such an event or todo to fail.  This commit ensures that the format is
explicitly set.

Task-number: MOBILITY-1931
Reviewed-by: Kevin Wu Won