Added initial mouse handling to Maps QML.
This commit adds a Maps specific mouse area and mouse event, which are
coordinate aware.

There is also a new QGeoMapObject class used so that the various
map objects can be proper QDeclarativeItems.

The declarative mapviewer example prints to the console when
various mouse events are inflicted on the circle object.

Still need to handle mouse clicks, press and hold, acceptance of
events, mouse grabbing and possibly dragging.