Generic handling of WAVE files
Changed to WAVE decoder to follow the file specification which states
that "Programs should expect (and ignore) any unknown chunks
encountered. However, fmt chunk must always occur before data chunk, and
both of these chunks are mandatory in a WAVE file."
This implementation uses a state variable to keep track of which chunk
we are currently looking for, and only reads when enough bytes are
available.

Task-number: QTMOBILITY-817
Reviewed-by: Andrew den Exter