Fixes: MOBILITY-2118
The tst_QLandmarkFileHandle_Lmx auto-test was failing
on harmattan because it was not using the sparql manager.

Also remove superfluous code of saving the categories
in order to given them ids becaue this is not necessary.

Also save landmark data in the datastore before any tests
are run and restore them afterwards.