Fix signal emission in Immersion feedback backend
Previously, if the state of the effect was queried prior to the
backend's internal timer for the effect could fire, but after
the effect was finished, no signal would be emitted for the state
change (to QFeedbackEffect::Stopped).  This resulted in a unit test
failure on platforms using the Immersion backend.

This change ensures that the function which is called when the state
of the effect is queried emits the signal manually (if required).