Corrected git hook complaints (tabs and whitespace).
Commits were old (separate dev repo), hence
opening the commits did not seem an option -> fixed with this
commit.  Furthermore few old commit messages do not have an empty
line in the message. I tried to fix this by rebasing but it got into
utterly mess because of merge conflicts --> let them be for now,
(and of course have all new commits comply :).