Use thumbnail instead of avatar in samplephonebook
Currently, the samplephonebook application attempts to use either
the thumbnail detail or the avatar detail, of the contact.
This can lead to a strange user experience where setting the avatar
has no outward effect on the application, because the thumbnail (which
is what is displayed) is not updated as a result of the operation.

This commit changes the samplephonebook so that it uses the thumbnail
detail only.

Task-number: MOBILITY-2002
Reviewed-by: Kevin Wu Won
Change-Id: I80f0498d455952087204600c2e5625bd1a120738