Stabilization of auto-test
Test was failing because, the CI system VM was unexpectedly fast.
We were trying to test cancel of the import operation however
the import operatio was already completed.  This resulted in the
QVERIFY macro failing and consequently the thread was deleted
before it was finished causing a segfault.

Task-number: MOBILITY-2188