Update the tap sensor for directionless detection.
Apparently some Symbian devices need this. The hardware
will report a tap event but it does not have a direction.

Tweak the definition of the tapDirection enum so that when
checked as a flag, things will work.

Don't use X, Y and Z for directionless tapping (these are for
detecting the axis without caring about the direction).
Instead, use new _Both flags in the Symbian backend.

Update the unit test so it checks the valid enum values.

Fixes: MOBILITY-1878