Cleaned up the handling of the program options.
This was initially just meant to add license headers, but when I
saw a global variable was being used I decided to start cleaning that
part of the example up a bit.

The whole example is really, really substandard from a Qt perspective
and will probably need to be removed or redone before long.