Turned some of the examples into system tests.
The geoservicedemo and mapviewer examples are very poor as far
as Qt examples go, which are meant to demonstrate both
proper usage of the API and good Qt style.

They have been used by a number of people as system tests and
then modified as such to the point that they're basically part
of the test infrastructure and we can't get rid of them.

Those examples are now being moved to tests/location-testing-tools
so that they can be used as system tests without polluting the
examples.

The new example for the Maps and Navigation API is mapsnavigator.
If you're going to make changes to the mapsnavigator example it's
probably best to talk to someone on the Qt Mobility team about it
first, because I'm not going to be shy about reverting suspicious
looking changes.