Avoid accidental usage of system includes for versit, versitorganizer.
These modules would use the QtContacts, QtVersit and QtOrganizer headers
already installed on the system, if any were present. That would lead
to compile errors if the headers were not compatible with those in the
source - and possibly other, more subtle errors.

Re-order INCLUDEPATH to ensure the headers from the source tree always
come first.

Reviewed-by: Chris Adams

Change-Id: If9cb1ea75fc7761331398b8b0a77d668f5158deb