Avoid accidental usage of system libraries.
While building Mobility, we were adding the Mobility lib install path to
the link path.  That's wrong, since we haven't installed Mobility yet.

On most platforms, this has not caused problems.
On Harmattan, it is quite likely that there is already some Mobility
libraries installed, which means that we would incorrectly link against
those instead of the ones we're building.

Reviewed-by: Alex