Fixed INCLUDEPATH problems for Mobility examples.
While building Mobility examples, we were adding the Mobility
headers install path to INCLUDEPATH.  That's wrong, since we haven't
installed Mobility yet.

On most platforms, this has not caused problems.
On Harmattan, it is quite likely that there are already some Mobility
headers installed, which means that we would incorrectly use those
instead of the ones from the Mobility we're building.

This is similar to the problems fixed by
e81921f362f76c83b4107ec429f5d4ad9e38ed7d and
31e1e2eaf6194c99b5a7a00691798907a89dd750.

Reviewed-by: Alex