Removed the handling of mouse movement from Maps QML bindings.
This includes both hovering and movement while the mouse
buttons are pressed.

There is insufficient time between now and the freeze for the
beta release to get the bindings for this functionality to a
level of quality which we are happy with.

They will be further developed and added when they are
completed to a high level of quality.